🚚 rename `triggerer` -> 'trigger'
