+++ bundles/conditional-availability-product-page-search/src/FondOfImpala/Zed/ConditionalAvailabilityProductPageSearch/Business/ConditionalAvailabilityProductPageSearchBusinessFactory.php
@@ -10,8 +10,8 @@ use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Generator
 use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Generator\StockStatusGeneratorInterface;
 use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Reader\ProductAbstractReader;
 use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Reader\ProductAbstractReaderInterface;
-use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Triggerer\StockStatusTriggerer;
-use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Triggerer\StockStatusTriggererInterface;
+use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Trigger\StockStatusTrigger;
+use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Trigger\StockStatusTriggerInterface;
 use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\ConditionalAvailabilityProductPageSearchDependencyProvider;
 use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Dependency\Facade\ConditionalAvailabilityProductPageSearchToConditionalAvailabilityFacadeInterface;
 use FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Dependency\Facade\ConditionalAvailabilityProductPageSearchToProductFacadeInterface;
@@ -63,11 +63,11 @@ class ConditionalAvailabilityProductPageSearchBusinessFactory extends AbstractBu
     }
 
     /**
-     * @return \FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Triggerer\StockStatusTriggererInterface
+     * @return \FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Trigger\StockStatusTriggerInterface
      */
-    public function createStockStatusTriggerer(): StockStatusTriggererInterface
+    public function createStockStatusTrigger(): StockStatusTriggerInterface
     {
-        return new StockStatusTriggerer(
+        return new StockStatusTrigger(
             $this->createProductAbstractReader(),
             $this->getProductPageSearchFacade(),
             $this->getRepository(),
