+++ bundles/conditional-availability-product-page-search/src/FondOfImpala/Zed/ConditionalAvailabilityProductPageSearch/Business/Trigger/StockStatusTriggerInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Trigger;
+
+interface StockStatusTriggerInterface
+{
+    /**
+     * @return void
+     */
+    public function trigger(): void;
+}
