--- bundles/conditional-availability-product-page-search/src/FondOfImpala/Zed/ConditionalAvailabilityProductPageSearch/Business/Triggerer/StockStatusTriggererInterface.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace FondOfImpala\Zed\ConditionalAvailabilityProductPageSearch\Business\Triggerer;
-
-interface StockStatusTriggererInterface
-{
-    /**
-     * @return void
-     */
-    public function trigger(): void;
-}
